update user table

+++ backend/migrations/2022-07-03-123945_create_users/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "user"
 (
-    id         INT GENERATED ALWAYS AS IDENTITY NOT NULL PRIMARY KEY,
+    id         BIGINT GENERATED ALWAYS AS IDENTITY NOT NULL PRIMARY KEY,
     login      TEXT                             NOT NULL,
     email      TEXT                             NOT NULL,
     created_at TIMESTAMP                        NOT NULL
